Removes current destination for a shuttle in transit from the destination list (#75170)
## About The Pull Request

Shuttles with multiple destinations will no longer show the current
destination as a selectable choice during transit.

<details>
<summary>Example</summary>
A shuttle has two possible destinations. A and B. <br>
While at A, the only valid destination is B and vice-versa. <br>
But while in transit to B, _both_ A and B can be chosen as destinations.
<br><br>
The change now hides B as a valid destination, as you are already going
there. <br>
It has the nice side effect of making shuttles with only two
destinations never showing a dropdown list while in transit.
</details>

## Why It's Good For The Game

Telling the shuttle to go to the same place as you're going to doesn't
do anything.

## Changelog
:cl:
qol: Shuttles will no longer show their current destination as a choice
while in transit.
/:cl: