updates martian can sprites, fixes misc text and bugs in wellcheers (#77757)
this is kind of a miscellaneous pile of fixes for soda.dmi, rolled up
with a larger resprite of the new cans that were added with #75988. you
can see everything visually changed (other than inhands) in the image
below

![image](https://github.com/tgstation/tgstation/assets/110322848/4c28f0c7-3975-4510-9958-9eef7f913051)
and then the inhands here:

![image](https://github.com/tgstation/tgstation/assets/110322848/03967e6e-6551-40d5-b2ec-7f27fabbc175)
also changes wellcheers mood event text, and the mood switch to be based
on sanity instead. apparently the mood test allowed for foolproof
traitor testing, which is not ideal. this hopefully fixes that.

the new martian drinks used the previous style of cans, which was
updated in #77424. this makes them consistent with that style.
everything else is minor things i missed or forgot to do in that PR.
also things that enable traitor testing are not great, even if they're
niche and require amputation to pull off.

:cl:
fix: wellcheers is now based on sanity rather than mood
image: resprites martian cans
image: adds inhands for martian cans and wellcheers
image: minor tweaks to other cans
spellcheck: updates wellcheers mood text
/:cl: