Fixes a bunch of callbacks that were being qdeleted, and code cleanup (#77904)
![image](https://github.com/tgstation/tgstation/assets/13398309/559eb50a-461c-4220-b628-55412baaffc3)

Continuing the work of
https://github.com/tgstation/tgstation/pull/77850.

it started with finding one that was being missed and causing a
runtime...then I noticed a whole lot more. While I was doing this I
found callbacks that weren't being nulled in `Destroy()`, so I added
that wherever I found these spots as well as some general code cleanup.

There were a lot more of these than I initially hoped to encounter so
I'm labeling it as a refactor.

Fixes lots of runtimes, improves code resiliency.

:cl:
refactor: fixed a bunch of instances of callbacks being qdeleted and
cleaned up related code
/:cl: