Fixes an issue with eyes not properly storing the old_eye_color var (#77692)
Somewhere down the line with all the refactors the code that saved the
previous eyes' color stopped working.

This just fixes it. It's used for things such as qol for the high
luminosity eyes. It can be considered a fix and QoL alike but I'm going
to label it as QoL.

![image](https://github.com/tgstation/tgstation/assets/13398309/74645c3a-92de-4ae0-86e4-f0d011549da0)

Restores previous functionality that was lost in a refactor

:cl:
qol: inserting new high luminosity eyes will now properly remember the
previous eyes' color and default to that initially
/:cl: