Holo para refund fix (#78431)
Holoparasite injector can be refunded if its been used, this is because
the uplink_reimburse element is not removed from it when its used. Added
some signal code to it so that this is fixed.

Fixes a bug.

Fixes a bug allowing holopara injectors to be refundable when used.

:cl:
fix: Fixes a bug allowing holopara injectors to be refundable when used.
/:cl: