Update 1, Fixes 21
Fixes some stuff.

Changes Engineering's layout to be closer to Paradise/TauCeti's style, should allow for a better redesign of Medical later on.

Parts of Medical were changed to allow for the Engineering changes.

The AI Satellite was changed to a design based on KiloStation's, this was done due to new space constraints.