The Big One (Update 7)
Makes major changes to Box's layout to better suit the switch from Bee-code to TG-code.

The majority of Service and all of Commons have been relocated to the west side. The Chapel and Library remain on the east side of the station, near departures.

Science has been relocated to be north of Medical.

Maintenance has been updated to fit with the area changes.

Arrivals has been updated to fit with the area changes.

Changed the Courtroom and other parts of Security slightly.

Updated the Random Bars and Engines to fit with the changes.

Slightly changed the External PA.

Adjusts the lights, again.

Maintenance is mostly filled, I've run out of ideas for now.

Changed the default emergency shuttle from Box to Cere in anticipation of higher populations.