Fixes and updates the TEG (#81044)
Fixes thermoelectric generators so they work again
Pipes can be connected to it, the machine can be rotated
Adds the ventcrawl movement flags as a defined bitfield
I also gave it a TGUI menu, it was small so it is pretty insignificant.
I added a little bit more text to error messages to make it clearer why
the thermoelectric generator isn't functional.
I also repathed generator to thermoelectric_generator because
'generator' is a keyword and is highlighted in green which makes people
using vscode a little confused what it's meant to be.

Old

![image](https://github.com/tgstation/tgstation/assets/53777086/0fe0651b-4d48-474b-869f-2f665757a0bc)

New

https://github.com/tgstation/tgstation/assets/53777086/064a5dda-5407-4817-b090-d22eb6c4aab8

This is one of the things I had to move to TGUI in
https://hackmd.io/XLt5MoRvRxuhFbwtk4VAUA
I was originally gonna remove it, but in the spirit of feature freeze I
thought I should at least give it a try.
This fixes many issues with it and gives it a new better UI that won't
stop updating easily so you can actually watch the changes as it
happens.
The TEG may not be obtainable in-game but it can still be mapped in or
give by admins, letting it function as intended is still a massive
benefit.

Closes https://github.com/tgstation/tgstation/issues/75738

:cl: JohnFulpWillard, Unit2E teaching me the TEG
fix: The TEG now works again (still unobtainable by regular means
though).
fix: the TEG and its circulators can now be rotated counterclockwise
again.
refactor: The TEG now uses a TGUI interface rather than the old HTML
one.
/:cl:

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>