BoxStation Post-Merge Fixes
Fixes the modular computers.

Fixes the appropriate request consoles to be able to make announcements again.

Adds the new machines to chemistry.

Adds the cargo shelves to the cargo bay and warehouse.

Moves the random bars and engines map files to where they need to be.

Slightly changes the layouts of Chemistry, Cryogenics, and the Exam Room.