Fixes pAIs permanently blocking requests (#82900)
Removes pAI subsystem/card's reliance on clients and makes them
unstoppable, which didn't really make much sense for it to be this way
in the first place since it's something that has to always happen and is
something that isn't client-controlled.

Also removed case of ``usr`` use in ui act.

Closes https://github.com/tgstation/tgstation/issues/70401

:cl:
fix: pAI requests should no longer randomly permanently break in a
round.
/:cl: