Fixed lights doing processing when they shouldn't need to. (#83313)
Lights would do processing if they spawned with a mock cell, since it
would never pass the conditional to see if the cell was full and stop
processing since the cell didn't exist. This fixes that by making it so
that having no cell equates to having a full one.

This is expected since whenever any charge is supposed to be used by the
cell, it will generate the cell which will have full charge.

Fixes a potential performance hog. Metastation dropped from 2800
processing machinery to 1900 after implementing this change.

No player facing changes

---------

Co-authored-by: Watermelon914 <3052169-Watermelon914@users.noreply.gitlab.com>